+++ graphdump/DtDump.h
@@ -2,6 +2,6 @@
 #include "../list.h"
 void DtStart (FILE* dotfile);
 void DtSetTitle(FILE* dotfile, List* lst);
-void DtSetNode (FILE* dotfile, List* lst, int i);
+void DtSetNode (FILE* dotfile, List* lst, long long i);
 void DtSetDependence (FILE* dotfile, List* lst);
 void DtEnd (FILE* dotfile);
